feat(cli): first implementation of CLI interactive shell
