+++ safe-cli/subcommands/cat.rs
@@ -8,7 +8,7 @@
 
 use super::helpers::get_from_arg_or_stdin;
 use super::OutputFmt;
-use crate::subcommands::auth::auth_connect;
+use crate::operations::auth::safe_connect;
 use log::debug;
 use prettytable::Table;
 use safe_api::{Safe, SafeData};
@@ -29,8 +29,8 @@ pub fn cat_commander(
     let url = get_from_arg_or_stdin(cmd.location, None)?;
     debug!("Running cat for: {:?}", &url);
 
-    // TODO: switch to connect_without_auth(safe)?;
-    auth_connect(safe)?;
+    // TODO: switch to safe_connect_without_auth(safe)?;
+    safe_connect(safe)?;
     let content = safe.fetch(&url)?;
     match &content {
         SafeData::FilesContainer {
