+++ safe-cli/subcommands/dog.rs
@@ -8,7 +8,7 @@
 
 use super::helpers::{get_from_arg_or_stdin, xorname_to_hex};
 use super::OutputFmt;
-use crate::subcommands::auth::auth_connect;
+use crate::operations::auth::safe_connect;
 use log::debug;
 use prettytable::Table;
 use safe_api::{NrsMapContainerInfo, Safe, SafeData};
@@ -28,9 +28,9 @@ pub fn dog_commander(
     let url = get_from_arg_or_stdin(cmd.location, None)?;
     debug!("Running dog for: {:?}", &url);
 
-    // TODO: switch to connect_without_auth(safe)?;
-    auth_connect(safe)?;
-    let content = safe.inspect(&url)?;
+    // TODO: switch to safe_connect_without_auth(safe)?;
+    safe_connect(safe)?;
+    let content = safe.fetch(&url)?;
     match &content {
         SafeData::FilesContainer {
             xorurl,
