+++ safe-cli/subcommands/fake_auth.rs
@@ -47,7 +47,7 @@ pub fn auth_connect(safe: &mut Safe) -> Result<(), String> {
 }
 
 #[allow(dead_code)]
-pub fn connect_without_auth(safe: &mut Safe) -> Result<(), String> {
+pub fn safe_connect_without_auth(safe: &mut Safe) -> Result<(), String> {
     debug!("Fake-auth is enabled so we don't try to read the credentials file");
     safe.connect(APP_ID, None)?;
